refactor(app): extract layout command registration
